software-map: move thread-target path builders to software-map-paths
Agent-Session: 01a067ce-34b7-7952-a023-ced44207d9ca
